refactor(twincat): extract shared constant-folding primitives (PR #1220)
Per garretfick's architectural note on PR #1220: xform_fold_
initializer_expressions imported try_fold_binary/try_fold_unary
(and, transitively, the arithmetic primitives they're built from)
directly from its sibling xform_fold_constant_expressions, creating
an implicit dependency of one xform_* pass on another's internals
with no real reason to exist.

Moves integer_value, make_integer_constant, make_real_constant,
fold_integer_binary, fold_real_binary, const_as_f64, try_fold_binary,
and try_fold_unary into a new pub(crate) module,
analyzer/src/constant_folding.rs. Both xform_fold_* passes now depend
on this shared leaf module instead of on each other.
